[shelf]: migrate shelves from one directory to another
* copy schemes and resources, then if everything ok -> delete old ones;
* check new directory permissions and validate related pathes files;
* execute migration as a Modal process, otherwise new appeared shelves
 can be lost;
* add an ability to cancel migration process and restore prev shelf dir;
* call Project.save and Application.saveSettings before
 dispose appropriate old shelf manager to be sure that
 all schemes were saved (clean old and write new);
* (IDEA-132033, IDEA-163428) shelve and migrate changes to custom dir;

(cherry picked from commit 3f664f9)

+++ platform/vcs-impl/src/com/intellij/openapi/vcs/changes/shelf/ShelveChangesManager.java
@@ -28,6 +28,7 @@ import com.intellij.openapi.diff.impl.patch.*;
 import com.intellij.openapi.diff.impl.patch.apply.ApplyFilePatchBase;
 import com.intellij.openapi.diff.impl.patch.formove.CustomBinaryPatchApplier;
 import com.intellij.openapi.diff.impl.patch.formove.PatchApplier;
+import com.intellij.openapi.options.ConfigurationException;
 import com.intellij.openapi.options.NonLazySchemeProcessor;
 import com.intellij.openapi.options.SchemeManager;
 import com.intellij.openapi.options.SchemeManagerFactory;
@@ -35,6 +36,7 @@ import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.progress.ProgressManager;
 import com.intellij.openapi.progress.Task;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.ui.Messages;
 import com.intellij.openapi.util.*;
 import com.intellij.openapi.util.io.FileUtil;
 import com.intellij.openapi.util.io.FileUtilRt;
@@ -77,6 +79,7 @@ import java.util.concurrent.TimeUnit;
 
 import static com.intellij.openapi.components.StoragePathMacros.PROJECT_CONFIG_DIR;
 import static com.intellij.openapi.vcs.changes.ChangeListUtil.getPredefinedChangeList;
+import static com.intellij.util.ObjectUtils.chooseNotNull;
 
 public class ShelveChangesManager extends AbstractProjectComponent implements JDOMExternalizable {
   private static final Logger LOG = Logger.getInstance("#com.intellij.openapi.vcs.changes.shelf.ShelveChangesManager");
@@ -86,7 +89,8 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
   @NonNls private static final String REMOVE_FILES_FROM_SHELF_STRATEGY = "remove_strategy";
 
   @NotNull private final PathMacroManager myPathMacroSubstitutor;
-  @NotNull private final SchemeManager<ShelvedChangeList> mySchemeManager;
+  @NotNull private SchemeManager<ShelvedChangeList> mySchemeManager;
+
   private ScheduledFuture<?> myCleaningFuture;
   private boolean myRemoveFilesFromShelf;
 
@@ -104,12 +108,14 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
   @NotNull
   public static String getDefaultShelfPresentationPath(@NotNull Project project) {
     //noinspection deprecation
-    return project.isDefault() ? PROJECT_CONFIG_DIR : Paths.get(getConfigDirPath(project), SHELVE_MANAGER_DIR_PATH).toString();
+    String projectConfig = project.isDefault() ? PROJECT_CONFIG_DIR : getConfigDirPath(project);
+    return Paths.get(projectConfig, SHELVE_MANAGER_DIR_PATH).toString();
   }
 
-  private static String getConfigDirPath(Project project) {
+  //returns system-dependant path
+  private static String getConfigDirPath(@NotNull Project project) {
     VirtualFile projectStoreDirectory = ProjectKt.getProjectStoreDirectory(project.getBaseDir());
-    return projectStoreDirectory != null ? projectStoreDirectory.getPath() : project.getBasePath();
+    return PathUtil.getLocalPath(chooseNotNull(projectStoreDirectory, project.getBaseDir()));
   }
 
   private final MessageBus myBus;
@@ -123,8 +129,7 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
     super(project);
     myPathMacroSubstitutor = PathMacroManager.getInstance(myProject);
     myBus = bus;
-    mySchemeManager =
-      createShelveSchemeManager(project, myPathMacroSubstitutor.expandPath(VcsConfiguration.getInstance(project).CUSTOM_SHELF_PATH));
+    mySchemeManager = createShelveSchemeManager(project, VcsConfiguration.getInstance(project).CUSTOM_SHELF_PATH);
 
     myCleaningFuture = JobScheduler.getScheduler().scheduleWithFixedDelay(() -> cleanSystemUnshelvedOlderOneWeek(), 1, 1, TimeUnit.DAYS);
     Disposer.register(project, new Disposable() {
@@ -152,7 +157,9 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
   @NotNull
   private SchemeManager<ShelvedChangeList> createShelveSchemeManager(@NotNull Project project,
                                                                      @Nullable String customPath) {
-    FilePath customShelfFilePath = customPath != null ? VcsUtil.getFilePath(customPath) : null;
+    FilePath customShelfFilePath = customPath != null ? VcsUtil.getFilePath(myPathMacroSubstitutor.expandPath(customPath)) : null;
+    //don't collapse custom paths
+    final boolean shouldCollapsePath = !VcsConfiguration.getInstance(myProject).USE_CUSTOM_SHELF_PATH;
     return SchemeManagerFactory.getInstance(project)
       .create(customShelfFilePath != null ? customShelfFilePath.getName() : SHELVE_MANAGER_DIR_PATH,
               new NonLazySchemeProcessor<ShelvedChangeList, ShelvedChangeList>() {
@@ -167,7 +174,9 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
                 public Parent writeScheme(@NotNull ShelvedChangeList scheme) throws WriteExternalException {
                   Element child = new Element(ELEMENT_CHANGELIST);
                   scheme.writeExternal(child);
-                  myPathMacroSubstitutor.collapsePaths(child);
+                  if (shouldCollapsePath) {
+                    myPathMacroSubstitutor.collapsePaths(child);
+                  }
                   return child;
                 }
               }, null, customPath != null ? Paths.get(customPath) : null);
@@ -195,6 +204,94 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
       return true;
     });
   }
+
+
+  public void checkAndMigrateUnderProgress(@Nullable String fromDirPath, @Nullable String toDirPath) throws ConfigurationException {
+    final SchemeManager<ShelvedChangeList> newSchemeManager = createShelveSchemeManager(myProject, toDirPath);
+    File fromFile = new File(getShelfPath(fromDirPath));
+    File toFile = new File(getShelfPath(toDirPath));
+    if (!fromFile.exists() || FileUtil.filesEqual(fromFile, toFile)) return; //previous shelf directory may do not exist at all
+    String validationError = validateDestinationDirectory(toFile);
+    if (validationError != null) throw new ConfigurationException(validationError);
+    newSchemeManager.loadSchemes();
+    if (VcsConfiguration.getInstance(myProject).MOVE_SHELVES) {
+      new Task.Modal(myProject, "Moving Shelves to the New Directory...", true) {
+        @Override
+        public void run(@NotNull ProgressIndicator indicator) {
+          for (ShelvedChangeList list : mySchemeManager.getAllSchemes()) {
+            if (!list.isValid()) continue;
+            ShelvedChangeList migratedList = ShelvedChangeList.copy(list);
+            //find new name;
+            File newTargetDirectory = suggestPatchName(myProject, migratedList.DESCRIPTION, toFile, "");
+            migrateResourcesTo(migratedList, newTargetDirectory, false);
+            newSchemeManager.addNewScheme(migratedList, false);
+            // migrate resources and scheme path
+            indicator.checkCanceled();
+          }
+          clearShelvedLists(mySchemeManager.getAllSchemes(), false);
+        }
+
+        @Override
+        public void onSuccess() {
+          super.onSuccess();
+          updateShelveSchemaManager(newSchemeManager);
+        }
+
+        @Override
+        public void onCancel() {
+          super.onCancel();
+          suggestToCancelMigrationOrRevertPathToPrevious();
+        }
+
+        private void suggestToCancelMigrationOrRevertPathToPrevious() {
+          if (Messages.showOkCancelDialog(myProject,
+                                          "Shelves moving failed. <br/>Would you like to use new shelf directory path or revert it to previous?",
+                                          "Shelf Error",
+                                          "&Use New",
+                                          "&Revert",
+                                          UIUtil.getWarningIcon()) == Messages.OK) {
+            updateShelveSchemaManager(newSchemeManager);
+          }
+          else {
+            VcsConfiguration vcsConfiguration = VcsConfiguration.getInstance(myProject);
+            vcsConfiguration.USE_CUSTOM_SHELF_PATH = fromDirPath != null;
+            vcsConfiguration.CUSTOM_SHELF_PATH = fromDirPath;
+          }
+        }
+
+        @Override
+        public void onThrowable(@NotNull Throwable error) {
+          super.onThrowable(error);
+          suggestToCancelMigrationOrRevertPathToPrevious();
+        }
+      }.queue();
+    }
+    else {
+      updateShelveSchemaManager(newSchemeManager);
+    }
+  }
+
+  private void updateShelveSchemaManager(SchemeManager<ShelvedChangeList> newSchemeManager) {
+    myProject.save();
+    ApplicationManager.getApplication().saveSettings();
+    SchemeManagerFactory.getInstance(myProject).dispose(mySchemeManager);
+    mySchemeManager = newSchemeManager;
+    notifyStateChanged();
+  }
+
+  @Nullable
+  public static String validateDestinationDirectory(@NotNull File destinationDir) {
+    if (!destinationDir.exists() && !destinationDir.mkdirs()) return "Can't find or create new shelf directory";
+    if (!destinationDir.canRead()) return "Destination shelf directory should have read access";
+    if (!destinationDir.canWrite()) return "Destination shelf directory should have write access";
+    return null;
+  }
+
+  @NotNull
+  //returns system-dependant path
+  private String getShelfPath(@Nullable String path) {
+    return chooseNotNull(path, getDefaultShelfPresentationPath(myProject));
+  }                          
 
   @NotNull
   public File getShelfResourcesDirectory() {
@@ -247,25 +344,30 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
   public Collection<String> checkAndMigrateOldPatchResourcesToNewSchemeStorage() {
     Collection<String> nonMigratedPaths = ContainerUtil.newArrayList();
     for (ShelvedChangeList list : mySchemeManager.getAllSchemes()) {
-      File patchDir = new File(getShelfResourcesDirectory(), list.getName());
-      nonMigratedPaths.addAll(migrateIfNeededToSchemeDir(list, patchDir));
+      File newPatchDir = new File(getShelfResourcesDirectory(), list.getName());
+      // it should be enough for migration to check if resource directory exists. If any bugs appeared add isAncestor checks for each path
+      if (!newPatchDir.exists() && newPatchDir.mkdirs()) {
+        nonMigratedPaths.addAll(migrateResourcesTo(list, newPatchDir, true));
+      }
     }
     return nonMigratedPaths;
   }
 
   @NotNull
-  private static Collection<String> migrateIfNeededToSchemeDir(@NotNull ShelvedChangeList list, @NotNull File targetDirectory) {
-    // it should be enough for migration to check if resource directory exists. If any bugs appeared add isAncestor checks for each path
-    if (targetDirectory.exists() || !targetDirectory.mkdirs()) return ContainerUtil.emptyList();
+  private static Collection<String> migrateResourcesTo(@NotNull ShelvedChangeList list,
+                                                       @NotNull File targetDirectory,
+                                                       boolean deleteOld) {
     Collection<String> nonMigratedPaths = ContainerUtil.newArrayList();
-    //try to move .patch file
+    //try to copy/move .patch file
     File patchFile = new File(list.PATH);
     if (patchFile.exists()) {
       File newPatchFile = getPatchFileInConfigDir(targetDirectory);
       try {
         FileUtil.copy(patchFile, newPatchFile);
         list.PATH = FileUtil.toSystemIndependentName(newPatchFile.getPath());
-        FileUtil.delete(patchFile);
+        if (deleteOld) {
+          FileUtil.delete(patchFile);
+        }
       }
       catch (IOException e) {
         nonMigratedPaths.add(list.PATH);
@@ -280,7 +382,9 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
           try {
             FileUtil.copy(shelvedFile, newShelvedFile);
             file.SHELVED_PATH = FileUtil.toSystemIndependentName(newShelvedFile.getPath());
-            FileUtil.delete(shelvedFile);
+            if (deleteOld) {
+              FileUtil.delete(shelvedFile);
+            }
           }
           catch (IOException e) {
             nonMigratedPaths.add(shelvedFile.getPath());
@@ -650,7 +754,7 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
                                                                                                            list.DATE.before(limitDate) &&
                                                                                                            (!onlyMarkedToDelete ||
                                                                                                             list.isMarkedToDelete()));
-    clearShelvedLists(toDelete);
+    clearShelvedLists(toDelete, true);
   }
 
   @CalledInAwt
@@ -867,16 +971,18 @@ public class ShelveChangesManager extends AbstractProjectComponent implements JD
   }
 
   public void clearRecycled() {
-    clearShelvedLists(getRecycledShelvedChangeLists());
+    clearShelvedLists(getRecycledShelvedChangeLists(), true);
   }
 
-  private void clearShelvedLists(@NotNull List<ShelvedChangeList> shelvedLists) {
+  private void clearShelvedLists(@NotNull List<ShelvedChangeList> shelvedLists, boolean updateView) {
     if (shelvedLists.isEmpty()) return;
     for (ShelvedChangeList list : shelvedLists) {
       deleteListImpl(list);
       mySchemeManager.removeScheme(list);
     }
-    notifyStateChanged();
+    if (updateView) {
+      notifyStateChanged();
+    }
   }
 
   private void recycleChangeList(@NotNull final ShelvedChangeList listCopy, @Nullable final ShelvedChangeList newList) {
